Clarify cluster network limitation
Since this page is about running a ceph cluster in a kubernetes cluster, I was initially confused by this statement.  Thinking that it was referring to network limitations of the kubernets cluster.  (public (node network) vs. private (overlay network)).

I know public and private  are the terms used for ceph, but since It initially confused me (someone who knows and works with K8s) it seemed like a good clarification to make.